update comenet codes for MD17/QM9/OC20
