+++ dig/threedgraph/method/comenet/features.py
@@ -2,12 +2,16 @@
 # https://github.com/TUM-DAML/gemnet_pytorch/blob/master/gemnet/model/layers/basis_utils.py
 # https://github.com/TUM-DAML/gemnet_pytorch/blob/master/gemnet/model/layers/basis_layers.py
 
-
+import math
 import torch
 import sympy as sym
 import numpy as np
 from scipy.optimize import brentq
 from scipy import special as sp
+
+from scipy.special import binom
+from torch_geometric.nn.models.schnet import GaussianSmearing
+
 
 def Jn(r, n):
     """
@@ -343,3 +347,138 @@ class TorsionEmbedding(torch.nn.Module):
         sbf = sbf.repeat_interleave(k, dim=1)
         out = rbf * sbf
         return out
+
+
+
+
+
+class SphericalBesselBasis(torch.nn.Module):
+    """
+    1D spherical Bessel basis
+
+    Parameters
+    ----------
+    num_radial: int
+        Controls maximum frequency.
+    cutoff: float
+        Cutoff distance in Angstrom.
+    """
+
+    def __init__(
+        self,
+        num_radial: int,
+        cutoff: float,
+    ):
+        super().__init__()
+        self.norm_const = math.sqrt(2 / cutoff)
+
+        # Initialize frequencies at canonical positions
+        self.frequencies = torch.nn.Parameter(
+            data=torch.tensor(
+                np.pi * np.arange(1, num_radial + 1, dtype=np.float32)
+            ),
+            requires_grad=True,
+        )
+
+    def forward(self, d_scaled):
+        return (
+            self.norm_const
+            / d_scaled[:, None]
+            * torch.sin(self.frequencies * d_scaled[:, None])
+        )  # (num_edges, num_radial)
+
+
+class BernsteinBasis(torch.nn.Module):
+    """
+    Bernstein polynomial basis,
+    as proposed in Unke, Chmiela, Gastegger, Schütt, Sauceda, Müller 2021.
+    SpookyNet: Learning Force Fields with Electronic Degrees of Freedom
+    and Nonlocal Effects
+
+    Parameters
+    ----------
+    num_radial: int
+        Controls maximum frequency.
+    pregamma_initial: float
+        Initial value of exponential coefficient gamma.
+        Default: gamma = 0.5 * a_0**-1 = 0.94486,
+        inverse softplus -> pregamma = log e**gamma - 1 = 0.45264
+    """
+
+    def __init__(
+        self,
+        num_radial: int,
+        pregamma_initial: float = 0.45264,
+    ):
+        super().__init__()
+        prefactor = binom(num_radial - 1, np.arange(num_radial))
+        self.register_buffer(
+            "prefactor",
+            torch.tensor(prefactor, dtype=torch.float),
+            persistent=False,
+        )
+
+        self.pregamma = torch.nn.Parameter(
+            data=torch.tensor(pregamma_initial, dtype=torch.float),
+            requires_grad=True,
+        )
+        self.softplus = torch.nn.Softplus()
+
+        exp1 = torch.arange(num_radial)
+        self.register_buffer("exp1", exp1[None, :], persistent=False)
+        exp2 = num_radial - 1 - exp1
+        self.register_buffer("exp2", exp2[None, :], persistent=False)
+
+    def forward(self, d_scaled):
+        gamma = self.softplus(self.pregamma)  # constrain to positive
+        exp_d = torch.exp(-gamma * d_scaled)[:, None]
+        return (
+            self.prefactor * (exp_d ** self.exp1) * ((1 - exp_d) ** self.exp2)
+        )
+
+
+class RadialBasis(torch.nn.Module):
+    """
+
+    Parameters
+    ----------
+    num_radial: int
+        Controls maximum frequency.
+    cutoff: float
+        Cutoff distance in Angstrom.
+    rbf: dict = {"name": "gaussian"}
+        Basis function and its hyperparameters.
+    envelope: dict = {"name": "polynomial", "exponent": 5}
+        Envelope function and its hyperparameters.
+    """
+
+    def __init__(
+        self,
+        num_radial: int,
+        cutoff: float,
+        rbf: dict = {"name": "gaussian"},
+    ):
+        super().__init__()
+        self.inv_cutoff = 1 / cutoff
+
+        rbf_name = rbf["name"].lower()
+        rbf_hparams = rbf.copy()
+        del rbf_hparams["name"]
+
+        # RBFs get distances scaled to be in [0, 1]
+        if rbf_name == "gaussian":
+            self.rbf = GaussianSmearing(
+                start=0, stop=1, num_gaussians=num_radial, **rbf_hparams
+            )
+        elif rbf_name == "spherical_bessel":
+            self.rbf = SphericalBesselBasis(
+                num_radial=num_radial, cutoff=cutoff, **rbf_hparams
+            )
+        elif rbf_name == "bernstein":
+            self.rbf = BernsteinBasis(num_radial=num_radial, **rbf_hparams)
+        else:
+            raise ValueError(f"Unknown radial basis function '{rbf_name}'.")
+
+    def forward(self, d):
+        d_scaled = d * self.inv_cutoff
+        return self.rbf(d_scaled)
